Move “success” and “failure” functions into the enum and make them static.
I prefer to keep helper and factory functions in the type they create. It keeps the namespace cleaner and makes it more obvious what they do.
It would be preferable to make these functions initializers, but Swift initializers aren't functions and can't be referenced and passed around.

I tried to do the same with the failure functions in Result+Foundation.swift, but the linker crashed for some reason.
I will place my attempts in a new branch and move on.